requirements.txt
datasets==2.21.0
openai==1.42.0
transformers==4.44.2
nltk==3.9.1
rouge_score==0.1.2
sentence-transformers==3.0.1
spacy==3.7.6
gensim==4.3.3
convokit==3.0.0
seaborn==0.13.2
sentencepiece==0.2.0
accelerate==0.33.0
azure-ai-inference==1.0.0b3
orjson==3.10.7